functions/requirements.txt
firebase_functions~=0.4.3
google-generativeai~=0.8.5
firebase-admin
Pillow~=11.3.0